+++ src/pages/Board/Write/BoardWrite.jsx
@@ -1,14 +1,14 @@
 import { Button, Editor, Input } from "@/components";
 import { BaseLayout } from "@/layouts";
 import { useAxios } from "@/hooks";
-import { useCallback, useEffect, useState } from "react";
+import { useCallback, useState } from "react";
 import styles from "./BoardWrite.module.css";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { boardWriteScheme } from "../constants";
 import { APIEndPoints } from "@/constants";
-import { uploadImageApi } from "../apis";
 import "quill/dist/quill.snow.css";
+import useQuillEvents from "@/hooks/useQuillEvents";
 
 const BoardWrite = () => {
   const [quill, setQuill] = useState(null);
@@ -45,59 +45,7 @@ const BoardWrite = () => {
     [accessToken, postBoard]
   );
 
-  /** Quill이 load되면 해당 Quill에 이벤트리스너를 달아줍니다 */
-  useEffect(() => {
-    if (!quill) return;
-    const editorElement = quill.root;
-    quill.on("text-change", () => {
-      setValue("content", quill.getContents());
-    });
-
-    editorElement.addEventListener("paste", (e) => {
-      const clipboardData = e.clipboardData;
-      if (clipboardData && clipboardData.items) {
-        for (let i = 0; i < clipboardData.items.length; i++) {
-          const item = clipboardData.items[i];
-          if (item.type.indexOf("image") !== -1) {
-            e.preventDefault();
-            e.stopImmediatePropagation();
-            const file = item.getAsFile();
-            handleImage(file);
-            break;
-          }
-        }
-      }
-    });
-
-    // 드래그앤드롭 이벤트 처리
-    editorElement.addEventListener("drop", (e) => {
-      const dt = e.dataTransfer;
-      if (dt && dt.files && dt.files.length) {
-        for (let i = 0; i < dt.files.length; i++) {
-          const file = dt.files[i];
-          if (file.type.indexOf("image") !== -1) {
-            handleImage(file);
-            return;
-          }
-        }
-      }
-    });
-
-    async function handleImage(file) {
-      const range = quill.getSelection();
-
-      try {
-        const {
-          data: { imageUrl },
-        } = await uploadImageApi(file);
-        quill.insertEmbed(range?.index ?? 0, "image", imageUrl, "user");
-      } catch (error) {
-        console.error("error", error);
-      }
-
-      quill.setSelection((range?.index ?? 0) + 1);
-    }
-  }, [accessToken, quill, setValue]);
+  useQuillEvents(quill, setValue);
 
   return (
     <BaseLayout>
